managed user jwt token in local storage and redirected user to home on successful login
