+++ src/utils/api_urls.js
@@ -1,0 +1,9 @@
+const BASE_PATH = "http://localhost:3000/";
+
+const APIS = {
+    FirebaseLoginAPI : BASE_PATH+"users/firebase-login",
+    SignUpAPI : BASE_PATH+"users/signup",
+    LoginAPI : BASE_PATH+"users/login"
+}
+
+export default APIS;
